saludapp/requirements.txt
appdirs==1.4.3
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
defusedxml==0.7.0rc1
distlib==0.3.0
Django==3.0.8
django-oauth-plus==2.2.9
django-tastypie==0.14.3
djangorestframework==3.9.0
filelock==3.0.12
Flask==1.1.2
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauth2==1.9.0.post1
oauthlib==3.1.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-mimeparse==1.6.0
python3-openid==3.2.0
pytz==2020.1
qrcode==6.1
requests==2.24.0
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
twilio==6.44.1
urllib3==1.25.10
virtualenv==20.0.20
Werkzeug==1.0.1
